requirements.txt
Django==1.11.4
Markdown==2.6.9
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
django-filter==1.0.4
django-map-widgets==0.1.9
djangorestframework==3.6.4
geocoder==1.32.1
idna==2.6
psycopg2==2.7.3
pytz==2017.2
ratelim==0.1.6
requests==2.18.4
six==1.11.0
urllib3==1.22
wsgiref==0.1.2
